+++ error_display/static/script.js
@@ -14,6 +14,10 @@ window.addEventListener("load", () => {
 
 	let heartbeatID = setInterval(fireHeartbeat, FREQ);
 
+	document.getElementById("archive-btn").addEventListener("click", () => {
+		fetch("/open_archive").catch(console.error);
+	});
+
 	// If the window is hidden (including when the Steam Overlay is closed!!), stop sending
 	// heartbeat messages so the server can die.
 	document.addEventListener("visibilitychange", () => {